change Venue to use Integer, dont need Long as an ID

+++ backend/src/main/java/com/code/theaterapp/venue/VenueController.java
@@ -20,7 +20,7 @@ public class VenueController {
     public List<VenueDTO> getAllVenues() { return venueService.getAllVenues(); }
 
     @GetMapping("/{id}")
-    public VenueDTO getVenue(@PathVariable("id") Long id) {
+    public VenueDTO getVenue(@PathVariable("id") Integer id) {
         return venueService.getVenueById(id);
     }
